release/V1.1.31

+++ aiurt-module-fault/aiurt-module-fault-api/src/main/java/com/aiurt/modules/fault/dto/FaultSystemMonthCountDTO.java
@@ -30,5 +30,5 @@ public class FaultSystemMonthCountDTO implements Serializable {
     private String systemName;
 
     @ApiModelProperty(value = "子系统维修次数")
-    private Integer frequency;
+    private Long frequency;
 }
